web-animations/animation-timeline/animation-timeline-play-paused-test.html

<!DOCTYPE html>
<html>
<head>
<title>Animation timeline play() method test. Test paused attribute initial value of newly-created AnimationPlayer</title>
<meta name="assert" content="Default value of the paused attribute is false">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#the-animationtimeline-interface">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var player = newAnimationPlayer(doc);

    assert_false(player.paused, 'Initial value of the paused attribute of the ' +
        'newly-created AnimationPlayer object should be false');
}, 'Test paused attribute initial value of newly-created AnimationPlayer');
</script>
</body>
</html>
